feat: add "Visit Default Project" button to the project list page (#1163)
