+++ frontend/src/layouts/ProjectLayout.vue
@@ -11,6 +11,12 @@
       {{ $t("project.mode.tenant") }}
     </span>
   </h1>
+  <BBAttention
+    v-if="project.id === DEFAULT_PROJECT_ID"
+    class="mx-6 mb-4"
+    :style="'INFO'"
+    :title="$t('project.overview.info-slot-content')"
+  />
   <BBTabFilter
     class="px-3 pb-2 border-b border-block-border"
     :responsive="false"
@@ -39,7 +45,7 @@ import { idFromSlug, isProjectOwner } from "../utils";
 import ArchiveBanner from "../components/ArchiveBanner.vue";
 import { BBTabFilterItem } from "../bbkit/types";
 import { useI18n } from "vue-i18n";
-import { Project } from "../types";
+import { Project, DEFAULT_PROJECT_ID } from "../types";
 import { useCurrentUser, useProjectStore } from "@/store";
 
 type ProjectTabItem = {
@@ -185,6 +191,7 @@ export default defineComponent({
     );
 
     return {
+      DEFAULT_PROJECT_ID,
       state,
       project,
       allowEdit,
